Renames interfaces and implementations for server socket and logger.
