+++ src/main/java/com/saralein/server/ShutdownHook.java
@@ -1,13 +1,13 @@
 package com.saralein.server;
 
-import com.saralein.server.logger.ILogger;
+import com.saralein.server.logger.Logger;
 import com.saralein.server.server.Server;
 
 public class ShutdownHook extends Thread {
     private final Server server;
-    private final ILogger logger;
+    private final Logger logger;
 
-    public ShutdownHook(Server server, ILogger logger) {
+    public ShutdownHook(Server server, Logger logger) {
         this.server = server;
         this.logger = logger;
     }
