+++ src/main/java/com/saralein/server/connection/ServerSocket.java
@@ -2,7 +2,7 @@ package com.saralein.server.connection;
 
 import java.io.IOException;
 
-public interface IServerSocket {
+public interface ServerSocket {
     int getPort();
     Connection accept() throws IOException;
     void close() throws IOException;
