+++ src/main/java/com/saralein/server/logger/ConnectionLogger.java
@@ -1,0 +1,15 @@
+package com.saralein.server.logger;
+
+import java.io.PrintStream;
+
+public class ConnectionLogger implements Logger {
+    private PrintStream printer;
+
+    public ConnectionLogger(PrintStream printer) {
+        this.printer = printer;
+    }
+
+    public void log(String status) {
+        printer.println(status);
+    }
+}
