--- src/main/java/com/saralein/server/logger/ILogger.java
@@ -1,5 +1,0 @@
-package com.saralein.server.logger;
-
-public interface ILogger {
-    void log(String status);
-}
