+++ src/test/java/com/saralein/server/logger/LoggerTest.java
@@ -11,7 +11,7 @@ public class LoggerTest {
     @Test
     public void loggerLogsStatusFromPrintStream() {
         ByteArrayOutputStream output = new ByteArrayOutputStream();
-        ILogger logger = new Logger(new PrintStream(output));
+        Logger logger = new ConnectionLogger(new PrintStream(output));
         logger.log("I am logging!");
         logger.log("I am still logging!");
         assertEquals("I am logging!\nI am still logging!\n", output.toString());
