+++ src/test/java/com/saralein/server/mocks/MockLogger.java
@@ -1,7 +1,7 @@
 package com.saralein.server.mocks;
 
-import com.saralein.server.logger.ILogger;
+import com.saralein.server.logger.Logger;
 
-public class MockLogger implements ILogger {
+public class MockLogger implements Logger {
     public void log(String status) {}
 }
