+++ src/test/java/com/saralein/server/mocks/MockServer.java
@@ -2,8 +2,6 @@ package com.saralein.server.mocks;
 
 import com.saralein.server.server.Server;
 
-import java.util.concurrent.ExecutorService;
-
 public class MockServer extends Server {
     private boolean stopCalled = false;
